Add function to check for project parameter
A restriction set in place to address issue #1.
Processing all subdirectories of src/ would be a nice feature
but presents far more issues.
Push the project focus and fail is no project is specified.
checkProj() checks if a project parameter was passed
and fails if not.
If a command is run with --force, the project value is set to ''.