Add 'clean' after archive copying
To complete the task list of issue #7.
At this time, archive does not clean the src/ folder
(in case of follow-up emails, etc. that will build on the same project).
If a mistake has been made, the files can always be built from src/
again, so requiring a successful copy isn't necessary to add at this time.